Move StakeStep to `@/accounts/components/`
