+++ packages/ui/src/council/modals/VoteForCouncil/VoteForCouncilFormModal.tsx
@@ -3,8 +3,8 @@ import React, { useCallback, useState } from 'react'
 import styled from 'styled-components'
 import { Event, EventData } from 'xstate/lib/types'
 
+import { StakeStep } from '@/accounts/components/StakeStep'
 import { ButtonPrimary } from '@/common/components/buttons'
-import { StakeStep } from '@/common/components/forms/StakeStep'
 import { Arrow } from '@/common/components/icons'
 import { Modal, ModalFooter, ModalHeader, ScrollableModalColumn, ScrolledModalBody } from '@/common/components/Modal'
 import { useModal } from '@/common/hooks/useModal'
